+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 Bug fixes
 - fix ang2pix bug w/ returning negative pixels (#45 #48)
+- add aliases healpix.RING and healpix.NEST for healpix.PixelOrder
 
 ## 2025.7.1